// frontend/src/generated/api/models/ArtifactPurpose.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EnergyForecast API
 * Energy consumption analysis and forecasting API
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Supported artifact purposes, persisted as ``app.artifacts.kind``.
 * @export
 */
export const ArtifactPurpose = {
    RawDataset: 'raw_dataset',
    Model: 'model',
    Metrics: 'metrics',
    Predictions: 'predictions',
    ForecastExport: 'forecast_export',
    Chart: 'chart',
    Manifest: 'manifest',
    Other: 'other'
} as const;
export type ArtifactPurpose = typeof ArtifactPurpose[keyof typeof ArtifactPurpose];


export function instanceOfArtifactPurpose(value: any): boolean {
    for (const key in ArtifactPurpose) {
        if (Object.prototype.hasOwnProperty.call(ArtifactPurpose, key)) {
            if (ArtifactPurpose[key as keyof typeof ArtifactPurpose] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ArtifactPurposeFromJSON(json: any): ArtifactPurpose {
    return ArtifactPurposeFromJSONTyped(json, false);
}

export function ArtifactPurposeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ArtifactPurpose {
    return json as ArtifactPurpose;
}

export function ArtifactPurposeToJSON(value?: ArtifactPurpose | null): any {
    return value as any;
}

export function ArtifactPurposeToJSONTyped(value: any, ignoreDiscriminator: boolean): ArtifactPurpose {
    return value as ArtifactPurpose;
}