// frontend/src/generated/api/models/ExportArtifactResponse.ts
export interface ExportArtifactResponse {
    /**
     * 
     * @type {Date}
     * @memberof ExportArtifactResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ExportArtifactResponse
     */
    downloadUrl: string;
    /**
     * 
     * @type {string}
     * @memberof ExportArtifactResponse
     */
    filename: string;
    /**
     * 
     * @type {string}
     * @memberof ExportArtifactResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ExportArtifactResponse
     */
    mediaType: string;
    /**
     * 
     * @type {ArtifactPurpose}
     * @memberof ExportArtifactResponse
     */
    purpose: ArtifactPurpose;
    /**
     * 
     * @type {string}
     * @memberof ExportArtifactResponse
     */
    sha256: string;
    /**
     * 
     * @type {number}
     * @memberof ExportArtifactResponse
     */
    sizeBytes: number;
}